Fix some issues with Series III.
The packet data in some records has an FCS and in some records doesn't.
It appears that only the Series III packets do.  Handle that.

Handle HT and VHT information supplied by Series III.

Get rid of an unused include while we're at it, and make the checks for
whether the protocol tree is being built a bit clearer, by testing the
tree pointer we'll actually be using.  Make some style cleanups.  Get
rid of some unused cruft.

Fix FPGA version checks.

This is based on Ixia's patches to an older version of Wireshark, but is
much cleaned up.

Bug: 11464

Change-Id: Ia341e6ffb1771cf38be812bf786f59b3250b7d5b
Reviewed-on: https://code.wireshark.org/review/10756
Reviewed-by: Guy Harris <guy@alum.mit.edu>