make-version.pl: no need for git-ls-remote
In subversion, the URL was needed to find the branch. This is not needed
for git where a 'git branch' command suffices.

Remove the unused URL handling from git, this fixes a warning during
cmake on RHEL6 and avoids a fallback to SVN:

    Version configuration file version.conf not found.  Using defaults.
    usage: git ls-remote [--heads] [--tags]  [-u <exec> | --upload-pack <exec>] <repository> <refs>...
    Unable to open ./.svn/entries
    #define VCSVERSION "v2.1.0rc0-69-g2eb7e87"
    #define VCSBRANCH "master"

Print the SVN warning to stderr, this avoids a compilation failure due
to a syntax error in version.h.

Change-Id: Idb28e74b7a9bf592686b45349cf5e81f85f6c358
Reviewed-on: https://code.wireshark.org/review/11041
Reviewed-by: Gerald Combs <gerald@wireshark.org>
(cherry picked from commit 1868dce4dee2ccfbb5ab3a60f88c192508b8ea13)
Reviewed-on: https://code.wireshark.org/review/14957
Reviewed-by: Guy Harris <guy@alum.mit.edu>