Treat "-" as "standard input" in the CLI, not in libwiretap.
That's a UI convention, and the GUI shouldn't honor that convention - a
user might get confused if they try to save to "-" and end up with
nothing (and with a ton of crap in a log file if programs launched from
the GUI end up with their standard output and error logged).

While we're at it, make randcap report write and close errors.

Change-Id: I6ea0a486864da71ecbb98501c96b98cd3bc424cb
Reviewed-on: https://code.wireshark.org/review/11667
Reviewed-by: Guy Harris <guy@alum.mit.edu>