Skip only the actual file descriptor close when writing to stdout.
Have a "this is stdout" flag for a wtap_dumper, and have "open the
standard output for dumping" routines that set that flag.  When closing
a wtap_dumper, do most of the work regardless of whether we're writing
to the standard output or not (so that everything gets written out) and
only skip the closing of the underlying file descriptor.

Change-Id: I80176dad0d9b5dff1b64c7d4703e0ab36501f547
Reviewed-on: https://code.wireshark.org/review/11674
Reviewed-by: Guy Harris <guy@alum.mit.edu>