Don't include wireshark-gtk.desktop in RPMs.
The RPMs use 'alternatives' to determine which GUI is used so it doesn't make
sense to have 2 desktop entries: one for 'wireshark' (Qt or Gtk GUI, depending on
configuration) and one for 'wireshark-gtk' (the Gtk GUI).

(Maybe it makes sense to just not use 'alternatives' and allow the user to pick
which GUI is used via the menu system.  But then if they wanted to run the Gtk+
GUI from the command line they'd need to remember to run 'wireshark-gtk' even
if that's the only GUI installed...)

Change-Id: I9d3fe13bb01eab87caad4ad21c6571ef6288b110
Reviewed-on: https://code.wireshark.org/review/11780
Reviewed-by: Balint Reczey <balint@balintreczey.hu>
Reviewed-by: Jeff Morriss <jeff.morriss.ws@gmail.com>
(cherry picked from commit fc51af45e65b534fa52c5519e23a7250ada195be)
Reviewed-on: https://code.wireshark.org/review/11873