Qt: Disable setStretchLastSection in the packet list.
QTreeView sets the stretchLastSection property of its header by default.
In our case this means that if the sum of our recent column widths
exceeds the width of the packet list viewport QHeaderView will shrink
the last column to fit.

Disable setStretchLastSection. We want its behavior when our columns are
too narrow so check for that in ::showEvent and temporarily enable it
there.

Bug: 11738
Change-Id: Ia4aad63e4f4bf899891bcebb7032dc5ebeb74cc7
Reviewed-on: https://code.wireshark.org/review/12392
Reviewed-by: Gerald Combs <gerald@wireshark.org>
Petri-Dish: Gerald Combs <gerald@wireshark.org>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>