Qt: No path for created and renamed profiles
The path is not valid for created and renamed profiles because the
directory is not created yet.  Use this label as a info label to
describe where the profile is created from or renamed from instead.

Use correct path for the Default profile.
Removed Bold from the current profile in status bar list.
Added tooltip for system provided profiles.

Change-Id: I61c8b1cc811dd9f9419ff9e373a8d00aa4e30446
Ping-Bug: 11704
Reviewed-on: https://code.wireshark.org/review/12447
Reviewed-by: Stig Bjørlykke <stig@bjorlykke.org>