Rename some variables to make it a bit clearer what they are.
rec_length_remaining is the amount of data we haven't already read from
the record; it starts out as the record length and gets decreased.  It
is not the length of data in the packet.

Change-Id: I46cd78e29aee13a686f1f6c8efbe258277e15686
Reviewed-on: https://code.wireshark.org/review/12759
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 1a5ed10bad99a96389d0ccbee5d7804c8cf98a0a)
Reviewed-on: https://code.wireshark.org/review/12760