Save the results of wmem_strndup(NULL, ...) in a non-const pointer.
That way, we have a non-const pointer to use when freeing it, and don't
have to undo the constification with a cast.

Rename "has_slash" to "slash", while we're at it, as it's not a Boolean
indicating whether the string has a slash, it's either a pointer to the
slash in question or NULL if the string has no slash.

Change-Id: Ia55b39bddb67c8ca71f7b09ee5eb82efaa3bdf0c
Reviewed-on: https://code.wireshark.org/review/12891
Reviewed-by: Guy Harris <guy@alum.mit.edu>